+++ src/filters/shaders/removeColor.shaders.ts
